requirements.txt
google-cloud-bigquery>=3.12.0
google-cloud-logging>=3.8.0
google-cloud-storage>=2.11.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
requests>=2.31.0
tenacity>=8.2.3
pytest>=7.4.2
responses>=0.23.3
pydantic>=2.4.2
pydantic-settings>=2.0.3
python-dotenv>=1.0.0
